+++ cstar/tests/unit_tests/orchestration/test_orchestration.py
@@ -12,7 +12,7 @@ from cstar.base.feature import (
     ENV_FF_ORCH_TRX_OVERRIDE,
     ENV_FF_ORCH_TRX_TIMESPLIT,
 )
-from cstar.base.utils import FLAG_ON
+from cstar.base.env import FLAG_ON
 from cstar.orchestration.launch.local import LocalLauncher
 from cstar.orchestration.models import Application, RomsMarblBlueprint, Step, Workplan
 from cstar.orchestration.orchestration import (
