+++ cstar/tests/unit_tests/system/test_environment.py
@@ -6,13 +6,8 @@ from unittest.mock import Mock, PropertyMock, call, mock_open, patch
 
 import pytest
 
-from cstar.base.utils import (
-    ENV_CSTAR_CACHE_HOME,
-    ENV_CSTAR_CONFIG_HOME,
-    ENV_CSTAR_DATA_HOME,
-    ENV_CSTAR_STATE_HOME,
-    get_env_item,
-)
+from cstar.base.env import get_env_item, ENV_CSTAR_CACHE_HOME, ENV_CSTAR_CONFIG_HOME, ENV_CSTAR_DATA_HOME, \
+    ENV_CSTAR_STATE_HOME
 from cstar.system.environment import CStarEnvironment
